+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ## 0.3.0 (upcoming)
 
-* Pending changelog
+* [EOS-1038] Output marathon-lb-sec logs to stdout
+* [EOS-987] Marathon-lb-sec logging format should comply with standard
+* Updated kms_utils version to 0.4.0
+* Included b-log version 0.4.0
+* Updated Marathon-LB main version v1.11.3
+* Bug fixing with race conditions
 
 ## 0.2.0 (December 19, 2017)